refactor: split construct module
